% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{document}
\alias{document}
\title{wrapper to devtools::document functions
with fluke project as context.}
\usage{
document()
}
\description{
wrapper to devtools::document functions
with fluke project as context.
}
