% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_project_config}
\alias{get_project_config}
\title{Get project metadata from \code{config/project.yaml}}
\usage{
get_project_config()
}
\description{
Get project metadata from \code{config/project.yaml}
}
