% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-targets.R
\name{get_targets}
\alias{get_targets}
\title{From a list of targets, filter specific targets
and get target names}
\usage{
get_targets(targets_list, regexp = NULL, names = FALSE)
}
\arguments{
\item{targets_list}{A list of targets}

\item{regexp}{Regular expression to filter list of targets.}

\item{names}{return names instead?}
}
\description{
From a list of targets, filter specific targets
and get target names
}
