% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{install}
\alias{install}
\title{Wrapper to devtools::install.
Installs the R package inside a fluke project.}
\usage{
install()
}
\description{
Wrapper to devtools::install.
Installs the R package inside a fluke project.
}
