% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{is_pipeline_dir}
\alias{is_pipeline_dir}
\title{identifies if a dirname is ane xisting pipeline path.}
\usage{
is_pipeline_dir(ppl)
}
\arguments{
\item{ppl}{A String. name of pipeline}
}
\value{
The full path of the pipeline.
}
\description{
identifies if a dirname is ane xisting pipeline path.
}
