% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_script_path}
\alias{pipeline_script_path}
\title{identifies if a dirname is an existing pipeline path.}
\usage{
pipeline_script_path(ppl)
}
\arguments{
\item{ppl}{A String. name of pipeline}
}
\value{
A list of the name of pipeline and  the full path of the pipeline.
}
\description{
identifies if a dirname is an existing pipeline path.
}
