% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{prepare_targets.pipelines}
\alias{prepare_targets.pipelines}
\title{Prepares pipelines for eventual \code{targets::tar_make}.}
\usage{
\method{prepare_targets}{pipelines}(pipelines, name = NULL)
}
\arguments{
\item{pipelines}{A pipelines object.}

\item{name}{name of a specific pipelines.}
}
\value{
A list of \code{tar_target} or \code{tar_target_raw}
objects from the \code{targets} package.
}
\description{
Prepares pipelines for eventual \code{targets::tar_make}.
}
