% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{python_env}
\alias{python_env}
\title{setup a python env to run reticulate.}
\usage{
python_env(...)
}
\arguments{
\item{...}{Additional argumentsa for \code{withr::with_envvar}}
}
\description{
setup a python env to run reticulate.
}
