% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{read a dataset using params from config.}
\usage{
read_dataset(path)
}
\arguments{
\item{dataset_path}{String. Path to dataset.}
}
\value{
A dataset object based on dataset type and \code{read_fun}.
}
\description{
read a dataset using params from config.
}
