% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{use_dataset}
\alias{use_dataset}
\title{Create a \code{tar_target} for using a dataset from \code{datasets/}}
\usage{
use_dataset(dataset)
}
\arguments{
\item{dataset}{A string or symbol.}
}
\description{
Create a \code{tar_target} for using a dataset from \code{datasets/}
}
